+++ .github/workflows/codestyle.yml
@@ -1,0 +1,15 @@
+name: Lint
+on:
+  - push
+  - pull_request
+jobs:
+  lint:
+    name: Lint
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: tonybajan/flake8-check-action@v1.0.0
+        with:
+          ignore: E722,W503
+          maxlinelength: 160
+          repotoken: ${{ secrets.GITHUB_TOKEN }}
