+++ requirements.txt
@@ -1,3 +1,8 @@
+matplotlib=3.1.3
+pandas==1.0.1
+dateparser==0.7.2
+discord.py==1.3.2
+scipy==1.4.1
 bleach==3.1.1
 certifi==2019.11.28
 chardet==3.0.4
